requirements.txt
flask
flask-cors
requests
docker
python-magic
yara-python
pefile
fpdf2
joblib
pandas
xgboost
